SmartOCR/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
python-multipart==0.0.9
anyio==4.2.0
pytest==8.0.0
httpx==0.26.0
pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
python-dotenv==1.0.1
pika==1.3.2
aio-pika==9.4.0
psycopg2-binary==2.9.9
minio==7.2.5
SQLAlchemy==2.0.25
opencv-python==4.9.0.80
